consolidated the Public and Private Key Child Derivation functions. also added new derivation test that doesnt use the node types
