+++ CardanoSharp.Wallet/Extensions/Models/PrivateKeyExtensions.cs
@@ -65,6 +65,11 @@ namespace CardanoSharp.Wallet.Extensions.Models
 
         public static PrivateKey Derive(this PrivateKey privateKey, string path)
         {
+            if (privateKey is null)
+            {
+                throw new ArgumentNullException(nameof(privateKey));
+            }
+
             if (!Bip32Utility.IsValidPath(path))
                 throw new FormatException("Invalid derivation path");
 
@@ -81,76 +86,10 @@ namespace CardanoSharp.Wallet.Extensions.Models
 
                 if (isHardened) index += MinHardIndex;
 
-                newPrivateKey = GetChildKeyDerivation(newPrivateKey, index);
+                newPrivateKey = Bip32Utility.GetChildKeyDerivation(newPrivateKey, index);
             }
 
             return newPrivateKey;
-        }
-
-        private static PrivateKey GetChildKeyDerivation(PrivateKey privateKey, uint index)
-        {
-            var kl = new byte[32];
-            Buffer.BlockCopy(privateKey.Key, 0, kl, 0, 32);
-            var kr = new byte[32];
-            Buffer.BlockCopy(privateKey.Key, 32, kr, 0, 32);
-
-            var z = new byte[64];
-            var zl = new byte[32];
-            var zr = new byte[32];
-            var i = new byte[64];
-            var seri = Bip32Utility.Le32(index);
-
-            BigEndianBuffer zBuffer = new BigEndianBuffer();
-            BigEndianBuffer iBuffer = new BigEndianBuffer();
-            if (Bip32Utility.FromIndex(index) == DerivationType.HARD)
-            {
-                zBuffer.Write(new byte[] { 0x00 }); //constant or enum?
-                zBuffer.Write(privateKey.Key);
-                zBuffer.Write(seri);
-
-                iBuffer.Write(new byte[] { 0x01 }); //constant or enum?
-                iBuffer.Write(privateKey.Key);
-                iBuffer.Write(seri);
-            }
-            else
-            {
-                var pk = privateKey.GetPublicKey(false);
-                zBuffer.Write(new byte[] { 0x02 }); //constant or enum?
-                zBuffer.Write(pk.Key);
-                zBuffer.Write(seri);
-
-                iBuffer.Write(new byte[] { 0x03 }); //constant or enum?
-                iBuffer.Write(pk.Key);
-                iBuffer.Write(seri);
-            }
-
-
-            using (HMACSHA512 hmacSha512 = new HMACSHA512(privateKey.Chaincode))
-            {
-                z = hmacSha512.ComputeHash(zBuffer.ToArray());
-                zl = z.Slice(0, 32);
-                zr = z.Slice(32);
-            }
-
-            // left = kl + 8 * trunc28(zl)
-            var left = Bip32Utility.Add28Mul8(kl, zl);
-            // right = zr + kr
-            var right = Bip32Utility.Add256Bits(kr, zr);
-
-            var key = new byte[left.Length + right.Length];
-            Buffer.BlockCopy(left, 0, key, 0, left.Length);
-            Buffer.BlockCopy(right, 0, key, left.Length, right.Length);
-
-            //chaincode
-
-            byte[] cc;
-            using (HMACSHA512 hmacSha512 = new HMACSHA512(privateKey.Chaincode))
-            {
-                i = hmacSha512.ComputeHash(iBuffer.ToArray());
-                cc = i.Slice(32);
-            }
-
-            return new PrivateKey(key, cc);
         }
     }
 }
